Initial project structure
Credit to https://dev.to/tendonnman/how-to-use-p5js-with-typescript-and-webpack-57ae for figuring out
the right incantations.

I had tried a more vanilla project structure, but couldn't get anything to work without webpack/parcel/etc.
I'd like to pair it down to just npm and ts, but was too challenging for where I'm at.